requirements.txt
# Core Platform Dependencies
fastapi==0.116.1
uvicorn==0.32.0
pydantic==2.8.0
pydantic-settings==2.2.1
python-dotenv==1.0.0

# Bot Framework Dependencies
python-telegram-bot==21.0
slack-sdk==3.26.1
aiohttp==3.9.1

# AI/ML Dependencies
openai==1.3.7
anthropic==0.7.8

# Infrastructure & Cloud Dependencies
boto3==1.34.0
google-cloud-storage==2.10.0
kubernetes==28.1.0
pyhelm==2.14.5

# Monitoring & Observability
prometheus-client==0.19.0
grafana-api==1.0.3
elasticsearch==8.11.0

# Data Processing & ML
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2

# Database & Storage
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
redis==5.0.1
pymongo==4.6.0
motor==3.3.2

# Testing Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
httpx==0.27.0

# Development Dependencies
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Security
cryptography==45.0.6
passlib==1.7.4
python-jose==3.3.0

# Utilities
click==8.1.7
rich==13.7.0
typer==0.9.0
structlog==23.2.0
